+++ backend/src/routes/fundingAgreements.js
@@ -3,15 +3,8 @@ const passport = require('passport')
 const router = express.Router()
 const auth = require('../components/auth')
 const isValidBackendToken = auth.isValidBackendToken()
-const {
-  getFundingAgreements,
-  updateFundingAgreement,
-  getFundingAgreementById,
-  getFundingPDFById,
-  getFundingReallocationRequests,
-  getFacilitiesWithExpiringOrRecentlyExpiredFAs,
-} = require('../components/fundingAgreements')
-const { param, query, body, validationResult, oneOf } = require('express-validator')
+const { getFundingAgreements, updateFundingAgreement, getFundingAgreementById, getFundingPDFById, getFundingReallocationRequests } = require('../components/fundingAgreements')
+const { param, query, validationResult, oneOf } = require('express-validator')
 const validateExpenseAuthority = require('../middlewares/validateExpenseAuthority.js')
 const validateFacility = require('../middlewares/validateFacility.js')
 const validatePermission = require('../middlewares/validatePermission.js')
@@ -97,17 +90,5 @@ router.get(
   (req, res) => {
     validationResult(req).throw()
     return getFundingReallocationRequests(req, res)
-  },
-)
-
-router.post(
-  '/facilities/expiring-or-recently-expired',
-  //passport.authenticate('jwt', { session: true }),
-  isValidBackendToken,
-  validatePermission(PERMISSIONS.VIEW_FUNDING_AGREEMENT),
-  [body('facilityIds', 'Request body [facilityIds] is required and must be an array').isArray({ min: 1 })],
-  (req, res) => {
-    validationResult(req).throw()
-    return getFacilitiesWithExpiringOrRecentlyExpiredFAs(req, res)
   },
 )
